Moves main-only utilities
Adds init message _after_ initial state is loaded
