+++ src/common/util/index.ts
@@ -1,4 +1,4 @@
-export * from "./mockStore";
-export * from "./environment";
 export * from "./display";
+export * from "./environment";
+export * from "./mockStore";
 export * from "./parsing";
